put ne_feature code in an explicit file
